sched: fair: avoid little cpus due to sync, prev_bias
Important threads can get forced to little cpu's
when the sync or prev_bias hints are followed
blindly. This patch adds a check to see whether
those paths are forcing the task to a cpu that
has less capacity than other cpu's available for
the task. If so, we ignore the sync and prev_bias
and allow the scheduler to make a free decision.

Bug: 110714207
Change-Id: Ie38061cd78a4e19c8783ffd460cc81caebfe9c75
Signed-off-by: John Dias <joaodias@google.com>

 Conflicts:
	kernel/sched/fair.c

[kras: Adjust for msm-4.9 after commit a7fc94a4e99e]
Signed-off-by: Albert I <kras@raphielgang.org>
Signed-off-by: improbir <ghosh_probir@yahoo.com>